+++ ui/src/main/java/io/snabble/sdk/ui/SnabbleUICallback.java
@@ -11,6 +11,8 @@ public interface SnabbleUICallback {
 
     void showSEPACardInput();
 
+    void showShoppingCart();
+
     void showPaymentCredentialsList();
 
     /**