+++ ui/src/main/java/io/snabble/sdk/ui/checkout/PaymentMethodView.java
@@ -231,10 +231,8 @@ class PaymentMethodView extends FrameLayout implements PaymentCredentialsStore.C
     }
 
     private void showPaymentNotPossibleDialog() {
-        // TODO i18n
         new AlertDialog.Builder(getContext())
-                .setTitle("Attention")
-                .setMessage("To add this payment method you need to secure your device using a screen lock (PIN, Fingerprint, ...) ")
+                .setMessage(R.string.Snabble_Keyguard_requireScreenLock)
                 .setPositiveButton(R.string.Snabble_OK, null)
                 .setCancelable(false)
                 .show();